+++ src/pages/Restaurants/RegionsContainer.jsx
@@ -4,9 +4,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './slice';
+} from '../../store/slice';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
